Fixed addressed code review feedbacks
